+++ lib/create-patch.sh
@@ -2,17 +2,17 @@
 # Use the unofficial bash strict mode: http://redsymbol.net/articles/unofficial-bash-strict-mode/
 set -euo pipefail; export FS=$'\n\t'
 
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+
 caskroom_dir=$1
 cask=$2
 branch=$3
 message=$4
 
-cd "${caskroom_dir}"
-echo "Linting cask ${cask}"
-brew cask style --fix "${cask}"
 echo "Auditing cask ${cask}"
-brew cask audit --download "${cask}"
+bash "${DIR}/audit.sh" "${cask}"
 echo "Creating branch for ${cask}"
+cd "${caskroom_dir}"
 git checkout -b "${branch}" --quiet
 git commit "${cask}.rb" --message "${message}" --quiet
 git push --force jcb "${branch}" --quiet